kvserver/rangefeed: wrap PerRangeEventSink inside BufferedPerRangeEventSink
This patch simplifies the BufferedStream and Stream interfaces by wrapping
*PerRangeEventSink inside BufferedPerRangeEventSink.

Epic: none
Release note: none
